crawl4AI-agent-unak/requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
python-dotenv
supabase
openai
gpt3-tokenizer
crawl4ai
